added some todos for future organisation improvements
